+++ lib/expo_sync/models/category_list.rb
@@ -3,8 +3,8 @@ module ExpoSync
     field :CategoryName, localize: true
     field :CategoryDescription, localize: true
 
-    def self.store_with_locale(data)
-      destroy_all
+    def self.store_with_locale(data, delta = false)
+      destroy_all if !delta
       grouped_data = data.group_by {|c| c[:CategoryID]}
       grouped_data.each do |id, category_locales|
         category = find_or_create_by(:CategoryID => id)
