+++ lib/expo_sync/models/contact_list.rb
@@ -7,8 +7,8 @@ module ExpoSync
     field :Position, localize: true
     field :ContactDescription, localize: true
 
-    def self.store_with_locale(data)
-      destroy_all
+    def self.store_with_locale(data, delta = false)
+      destroy_all if !delta
       grouped_data = data.group_by {|a| a[:ContactID]}
       grouped_data.each do |id, contact_locales|
         contact = find_or_create_by(:ContactID => id)
